Enhancement (task): Use literal here-strings for test and sample .ps1 definition object task action command arguments
Doing so eliminates pollution of task action arguments with escape characters allowing for improved reusability of commands which can be used straight on the command line or copied into separate scripts and vice-versa.

The change also, through the revision to use literal here-strings, includes fixes to relevant task action arguments used in sample and test definitions to be agnostic to the shell used.